It's a common thread to find the largest subsequence satisfying a certain condition.
It's almost always going to be the case that the most performant (not simplest) strategy is to find the left and right sections which together satisfy the negation of the condition. Then what's left is your subsequence!

Definitely need to work on having it be simpler and more idiomatic. Right now it looks messy. But it's just a matter of edge-casing clearly and handling details.